requirements.txt
asgiref==3.6.0
Django==4.2.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-seed==0.3.1
Faker==18.7.0
psycopg2==2.9.6
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.3
